Add support for targets that implement the C standard library
The purpose of this changes is to support Linux targets, but the code
should be suitable for many platforms - it just requires "printf" and
"exit".

It does not support tests that require system (or test process) resets
or watchdog support.

In the platform's Makefile, all of the PSA_*_IMPLEMENTED variables are
set to 0 because we don't know what PSA functionality will be available
on the target board. Users will have to modify these variables as
required.