Don't define 32-bit specific code if we don't need to
Some of the code in val_dispatcher.{h,c} deals with 32-bit ELF headers
and isn't suitable for 64-bit targets. GCC gives warnings/errors like
the following on 64-bit platforms:
```
./val/nspe/val_dispatcher.c: In function 'val_copy_elf':
./val/nspe/val_dispatcher.c:109:29: error: cast to pointer from integer of different size [-Werror=int-to-pointer-cast]
                             (uint8_t *)test_ph.p_paddr, test_ph.p_filesz))
                             ^
cc1: all warnings being treated as errors
```
We only actually need to define the ELF header code when
TEST_COMBINE_ARCHIVE == 0, so protect it with #ifdefs so that we can
build for 64-bit targets when TEST_COMBINE_ARCHIVE == 1.